requirements.txt
Django==4.0.3
cryptography==36.0.1
pymysql==1.0.2
djangorestframework==3.13.0
django-filter==21.1